init task at server startup (create keys in redis)
