+++ vertx/example-configs/config-depl.json
@@ -8,7 +8,6 @@
 		{
 			"id": "iudx.ingestion.pipeline.rabbitmq.RabbitMQVerticle",
 			"verticleInstances": 1,
-			
 			"dataBrokerIP": "",
 			"dataBrokerPort": 1234,
 			"dataBrokerVhost": "",
@@ -22,20 +21,23 @@
 			"networkRecoveryInterval": 500,
 			"automaticRecoveryEnabled": true
 		},
-                {
-                        "id": "iudx.ingestion.pipeline.redis.RedisVerticle",
-                        "verticleInstances": 1,
-                        "redisMode": "CLUSTER",
-                        "redisUsername": "xyz",
-                        "redisPassword": "abcd",
-                        "redisMaxWaitingHandlers": 5000,
-                        "redisHost":"redis-redis-cluster.redis.svc.cluster.local",
-                        "redisPort": 6379
-                },
-
+		{
+			"id": "iudx.ingestion.pipeline.redis.RedisVerticle",
+			"verticleInstances": 1,
+			"redisMode": "CLUSTER",
+			"redisUsername": "xyz",
+			"redisPassword": "abcd",
+			"redisMaxWaitingHandlers": 5000,
+			"redisHost": "redis-redis-cluster.redis.svc.cluster.local",
+			"redisPort": 6379
+		},
 		{
 			"id": "iudx.ingestion.pipeline.processor.ProcessorVerticle",
-			"verticleInstances": 1
+			"verticleInstances": 1,
+			"catalogueHost": "",
+			"cataloguePort": 443,
+			"keystore": "",
+			"keystorePassword": ""
 		}
 	]
 }
